requirements.txt
# Qwen3-4B (HF): generate.py, main.py, cpt_train.py
torch>=2.0
transformers>=4.51
accelerate>=0.26
peft>=0.10
pypdf>=4.0
huggingface_hub>=0.20
hf_transfer>=0.1
python-dotenv>=1.0
requests>=2.28
paramiko>=3.0
runpod-flash